% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDIS_plot_2D.R
\name{GDIS_plot_2D}
\alias{GDIS_plot_2D}
\title{Plot GDIS 2D plot}
\usage{
GDIS_plot_2D(
  input_triangle_parameters,
  x_lower = NULL,
  x_upper = NULL,
  y_lower = NULL,
  y_upper = NULL
)
}
\arguments{
\item{input_triangle_parameters}{the path to the RData GDIS output, which ends with 2D.triangle_parameters.RData}

\item{x_lower}{set lower value for the x axis}

\item{x_upper}{set upper value for the x axis}

\item{y_lower}{set lower value for the y asis}

\item{y_upper}{set upper value for the y axis}
}
\description{
This function plots a GDIS visualization based on GDIS parameters.
}
