% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GDIS_plot_CD.R
\name{GDIS_plot_CD}
\alias{GDIS_plot_CD}
\title{Plot GDIS cross-disorder plot}
\usage{
GDIS_plot_CD(
  input_triangle_parameters,
  show.rendering = TRUE,
  show.names = TRUE,
  webversion = FALSE
)
}
\arguments{
\item{input_triangle_parameters}{the path to the RData GDIS output, which ends with CD.triangle_parameters.RData}

\item{show.rendering}{option to show the rendering of the plot in an external window, default is TRUE}

\item{show.names}{option to show the names of the groups, default is TRUE}

\item{webversion}{needs to stay on FALSE}
}
\description{
This function plots a GDIS visualization based on GDIS parameters.
}
